modification work properly, display rating

+++ LITReview/delete/views.py
@@ -4,7 +4,7 @@ from django.db.models import Q
 from ask_review.models import Ticket
 from create_review.models import Review
 
-def delete(request, id_delete):
+def delete(request, content, id_delete):
     """[summary]
 
     Args:
@@ -14,7 +14,13 @@ def delete(request, id_delete):
         [type]: [description]
     """
     try:
-        Ticket.objects.get(Q(id=id_delete)&Q(user=request.user)) # .delete()
-        return redirect("/posts/")
+        if content == "ticket":
+            Ticket.objects.get(Q(id=id_delete)&Q(user=request.user)) # .delete()
+            return redirect("/posts/")
+        if content == "review":
+            Review.objects.get(Q(id=id_delete)&Q(user=request.user)) # .delete()
+            return redirect("/posts/")
     except Ticket.DoesNotExist:
+        return redirect("/flow/")
+    except Review.DoesNotExist:
         return redirect("/flow/")
